주문 조회 (사용자/관리자) , 삭제 기능 구현

See merge request sw_track/class_02_seoul/web_project/team7/shopping-mall-racer!61

+++ src/views/account/orders/orders.js
@@ -5,20 +5,25 @@ try {
   const tableList = document.querySelector('.account-page-body');
   console.log(res.result);
   if (res.result) {
-    for (let i = 0; i < res.result.length; i++) {
-      const orderDate = res.result[i].createdAt.substr(0, 10);
-      console.log(res.result);
-      const createTr = `
-          <div class="body-list body-tr">
-              <div class="date">${orderDate}</div>
-              <div class="product">${res.result[i].order_data} / ${res.result[i].quantity}개</div>
-              <div class="state">상품 준비중</div>
-              <div class="apply">
-                  <button class="button is-success" id="orderDeleteBtn" data-pid = ${res.result[i]._id}>주문취소</button>
-              </div>
-          </div>`;
-      tableList.insertAdjacentHTML('beforeend', createTr);
-    }
+    res.result.forEach(row => {
+      const orderRow = `
+      <div class="body-list body-tr">
+        <div class="createdAt">${row.createdAt.substr(0, 10)}</div>
+        <div class="fullName">${row.fullName}</div>
+        <div class="email">${row.email}</div>
+        <div class="phoneNumber">${row.phoneNumber}</div>
+        
+        <div class="order_data">${row.order_data}</div>
+        <div class="price">${row.price}</div>
+        <div class="request">${row.request}</div>
+        <button class="cancel button is-danger is-small" data-oid = ${
+          row._id
+        } id="orderDeleteBtn">주문 취소</button>
+        
+      </div>
+      `;
+      tableList.insertAdjacentHTML('beforeend', orderRow);
+    });
   } else {
     alert('주문 데이터가 없습니다.');
   }
@@ -29,19 +34,21 @@ try {
 const orderDeleteBtn = document.querySelectorAll('#orderDeleteBtn');
 orderDeleteBtn.forEach(btn => {
   btn.addEventListener('click', async e => {
-    // alert(e.target.dataset.pid);
-    const orderId = e.target.dataset.pid;
-    try {
-      const data = { orderId };
-      const res = await Api.delete('/api/order', data);
+    // alert(e.target.dataset.oid);
+    if (confirm('주문 취소 하시겠습니까?')) {
+      const orderId = e.target.dataset.oid;
+      try {
+        const data = { orderId };
+        const res = await Api.delete('/api/order', data);
 
-      alert(`${res.result.order_data} ${res.message} `);
-      // console.log(res);
-      // 로그인 페이지 이동
-      window.location.href = '/account/orders/orders.html';
-    } catch (err) {
-      alert(err);
-      // error 메세지 어떻게 따오지?
+        alert(`${res.result.order_data} ${res.message} `);
+        // console.log(res);
+        // 로그인 페이지 이동
+        window.location.href = '/account/orders/orders.html';
+      } catch (err) {
+        alert(err);
+        // error 메세지 어떻게 따오지?
+      }
     }
   });
 });
